refactor: dto, interceptor 위치 변경 및 @NoArgsConstructor에 AccessLevel 변경
- access = AccessLevel.PRIVATE로 변경
