+++ src/main/java/techcourse/myblog/config/WebMvcConfig.java
@@ -3,8 +3,8 @@ package techcourse.myblog.config;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import techcourse.myblog.interceptor.LoginInterceptor;
-import techcourse.myblog.interceptor.UserInterceptor;
+import techcourse.myblog.web.interceptor.LoginInterceptor;
+import techcourse.myblog.web.interceptor.UserInterceptor;
 
 @Configuration
 public class WebMvcConfig implements WebMvcConfigurer {
