+++ src/main/java/techcourse/myblog/web/LoginController.java
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import techcourse.myblog.domain.User;
 import techcourse.myblog.domain.UserRepository;
+import techcourse.myblog.web.dto.UserDto;
 
 import javax.servlet.http.HttpServletRequest;
 import java.util.Optional;
@@ -14,7 +15,7 @@ import java.util.Optional;
 @Slf4j
 @Controller
 public class LoginController {
-    UserRepository userRepository;
+    private final UserRepository userRepository;
 
     public LoginController(UserRepository userRepository) {
         this.userRepository = userRepository;
