+++ src/main/java/techcourse/myblog/web/dto/UserDto.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.web;
+package techcourse.myblog.web.dto;
 
 import lombok.Data;
 import techcourse.myblog.domain.User;
@@ -15,7 +15,6 @@ public class UserDto {
     @Pattern(regexp = "^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[~`!@#$%\\^&*()-]).{8,}$", message = "소문자, 대문자, 숫자, 특수문자가 조합된 8글자 이상을 입력하세요 ")
     private String password;
     @Email(message = "Wrong Email")
-    @Column(unique = true)
     private String email;
 
     public User toUser() {
